Create App folder and move relevant files to this folder
